Fixed Bug with user creation.
Bug in user creation caused by switching from bcrypt package
to bcrypt-nodejs. bcrypt-nodejs looks for an in progress function
AND a callback function.
