+++ db/index.js
@@ -1,4 +1,5 @@
 const mongoose = require('mongoose');
+const hash = require('./helper.js');
 mongoose.connect('mongodb://localhost/dinasour');
 
 const db = mongoose.connection;
@@ -11,8 +12,6 @@ db.once('open', function() {
   console.log('Mongoose connected');
 });
 
-const bcrypt = require('bcrypt-nodejs');
-const saltRounds = 10;
 
 //TODO: ADD USER SCHEMA
 let userSchema = mongoose.Schema({
@@ -30,7 +29,6 @@ let createUser = (user, callback) => {
   console.log(user);
   User.findOne({ email: user.email }, (err, existingUser) => {
     if (err) {
-      console.log(err);
       callback(err, null);
     }
     if (existingUser) {
@@ -41,26 +39,12 @@ let createUser = (user, callback) => {
         callback(null, { messageCode: 102, message: 'User name already exists' });
       }
     } else {
-      console.log('I got here');
-      bcrypt.hash(user.password, saltRounds, function(err, hash) {
-        if (err) {
-          callback(err, null);
-        }
-
-        let user_obj = {
-          firstName: user.firstName,
-          lastName: user.lastName,
-          userName: user.userName,
-          email: user.email,
-          password: hash
-        }; 
-
-        let newUser = new User(user_obj);
-        newUser.save()
-        .then(data => {
+      hash.hashPass(user, (err, userResult) => {
+        let newUser = new User(userResult);
+        console.log(userResult);
+        newUser.save().then(data => {
           callback(null, data)
-        })
-        .catch(data => {
+        }).catch(error => {
           callback(error, null)
         })
       });
