+++ server/login.js
@@ -1,6 +1,7 @@
 const db = require('../db/index.js');
 const app = require('./index.js');
 const login = require('express').Router();
+const util = require('./helpers/utilities.js');
 
 login.get('/login', function(req, res) {
   //render login page/modal
